Rename chart rke2-canal-v3.13.300-build20210223
This seems to be an issue with the regenerate-assets script, most likely introduced due to the weird naming scheme for this chart which was allowed in the old build scripts.

The new build scripts expects the chart version to always end with the packageVersion, e.g. this should be `rke2-canal-v3.13.3-build2021022300`, but since this is already released I'm not going to tamper with it.

Signed-off-by: Arvind Iyengar <arvind.iyengar@rancher.com>