Bump all packageVersions and introduce comment
Why do we need to bump all the packageVersions?

The new charts-build-scripts treat Chart.yaml and requirements.yaml as "managed files", which means that the scripts themselves can make changes to those files.

When the scripts touch those files (e.g. to add a new dependency based on the contents of generated-changes/dependencies or to patch a rcVersion or packageVersion to the chart), dumping back out the YAML results in slight changes due to the unmarshalling process, such as the re-ordering of annotations.

Since these minor changes comprise a change introduced by Rancher and the scripts refuse changes that modify already released packages, the simple fix for this is just to universally bump the packageVersion as part of the migration process.

As part of creating this PR, I generated these minor changes in https://github.com/aiyengar2/charts-diff/tree/rke2-charts/diff so that we can sign off that these are indeed inconsequential to the migration. For context, the process used to generate these files was:
- Pull in the current branch. Run make charts with the old scripts
- Pull in the migration branch. Run make charts with the new scripts
- Dump the diff between them in a patch file

Signed-off-by: Arvind Iyengar <arvind.iyengar@rancher.com>