Do not explicitly configure Calico MTU
This change removes explicit configuration of Calico MTU to let Calico auto-detect host MTU and use it for workload interfaces.

For context, Calico auto-detects host network interface MTU and uses it to configure the MTU for workload interfaces. When a non-zero MTU value is explicitly passed to Calico, it overrides the auto-detected MTU value. If the passed MTU value is greater than the host interface value, it could lead to fragmentation and packet loss.